Clean dashboard notices and split Group A tests
- Remove recovery-code and no-passkey notices from the dashboard
- Keep recovery-code status scoped to the MFA setup page
- Keep no-passkey state scoped to the security keys page
- Remove unused recovery-code dashboard calculation
- Add regression coverage for dashboard notice placement and masking
- Add registration, duplicate phone, mass-assignment, and DAST payload coverage
- Replace the monolithic Group A security test file with focused test modules
- Add shared auth-flow test helpers
- Preserve existing security and slow marker behavior for the split tests
- Update pytest optimization expectations for the new test inventory
