+++ app/static/js/dashboard.js
@@ -16,5 +16,4 @@
   }
 
   makeToggle('bal-eye-btn',  'card-balance-masked', 'card-balance-full', 'bal-eye-icon',  'Show balance',        'Hide balance');
-  makeToggle('acct-eye-btn', 'card-acct-masked',    'card-acct-full',    'acct-eye-icon', 'Show account number', 'Hide account number');
 }());
